LayoutTests/fast/events/before-input-events-prevent-recomposition-expected.txt

To manually test this, try to recompose text. The original text should not be modified.

initial composition text:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "foo"

after preventing recomposition:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "foofoo"

after selecting a new composition string:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "foogarply"

after committing the new composition string:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "foogarply"
